Improvements of pcap handling
When writing a Header to pcap, it should not be modified and thus be
passed as a constant.

This commits adds a sink to export Header along the packet. This is
done for performance reasons, to prevent the (costly) creation of a new packet with the header.

I've also modified some tests to remove unnecessary data copy via using
the overload that accepts a packet as a parameter.