Flow control Part I: NetDeviceQueues can be started, stopped and woken
Add public methods to the NetDeviceQueue class to start, stop or wake
a transmission queue. These methods are primarily called by NetDevices,
but can also be invoked by the network stack once techniques like BQL
are implemented. An IsStopped method is also added to allow queue discs to
check the current status of a transmission queue. A wake callback is provided,
so that upper layers (queue discs) can be solicited to send packets down.