NetDevices return the index of the transmission queue they select for a given packet
This commit just adds a GetSelectedQueue method to the NetDevice class. This
method is called by the Traffic Control layer before enqueuing a packet in the
queue disc to determine in which transmission queue the device will enqueue the
packet. Such information is exploited by multi-queue aware queue discs.
By default, this function returns 0, so that single-queue devices do not need
to be modified.