lte: Fix EPC test crashes by removing an assert
That's bad, usually the test must be reworked to not fail the assert.
However, there's no time in fixing a decade-old test, and also the code
worked like this since... forever. My fault was to add the assert in the first
instance, thinking that it would be a good service to the users. I was wrong.