wifi: The state of an OriginatorBlockAckAgreement is never set to INACTIVE
Hence, remove the INACTIVE state. Actually, there is no point
in having such a state. If the inactivity timeout expires, the
block ack agreement is torn down. If the originator decides to
switch to Normal Ack policy, there is no need for the block ack
agreement to change state ("The block ack record shall be updated
irrespective of the acknowledgment type (Normal or Block Ack) for
the TID with a block ack agreement" - Section 10.24.3 of 802.11-2016).