mobility: (merges !101) Add CartesianToGeographic conversion with test
There are cases where it is useful to return to geographic
coordinates from cartesian.  One example is when you have a
mobile node, the position is updated in cartesian coordinates
but you may have a need to understand the geographic position
or heading when determining position relative to other objects
placed geographically, or to compute compass heading or surface
speed.