core: (fixes 2815) Config::Connect should throw warnings
This patch logs unconditionally when a trace path doesn't exist.

We also provide FailSafe versions of Config::Connect and
Config::ConnectWithoutContext. These should be used
when you can't be sure that a trace path exists in the current model.

This resolves the warnings from AnimationInterface::ConnectCallbacks()

Credit to "InquisitivePenguin (GCI 2019) <>" for diagnosing this issue,
opening the way to a solution.