wifi: use uint16_t to count MPDUs in A-MPDU
Since IEEE 802.11ax standard increased maximum Block Ack Bitmap length
to 256, and HeConfiguration::MpduBufferSize can be set to value 256,
numbers of MPDUs in block ack manager and rate control should be
stored in uint16_t.

Before this patch, when 256 MPDUs were received in one A-MPDU,
nSuccessfulMpdus counter overflowed and was accounted as 0.