nix-vector-routing: Add unorderd_map Ptr<NetDevice> to Ptr<IpInterface>
This change improves the performance of Nix in bigger topologies as
it uses the `unordered_map` which has better performance than
`GetObject<Ip> ()` as it reads through the list.

This `unordered_map` will be created along with `IpAddress` to `Ptr<Node>`
`unordered_map`. Since it is a static member, it will be only time over
all the `NixVectorRouting` instances.

Signed-off-by: Ameya Deshpande <ameyanrd@outlook.com>