wifi: Add TX center frequency to WifiSpectrumSignalParameters
In this way, SpectrumWifiPhy::StartRx() does not have to obtain
the transmitting PHY just to retrieve the TX center frequency. Also,
the current approach does not work with multi-link devices because the
receiver should be able to know which PHY was used on the transmitter
side to send the frame.