Support the latest clang-format only
Motivation:
Different versions of clang-format tools which produce
different formatting outcomes. Testing the conformity of the different
formatting outcomes under a single criteria is an oxymoron. There is
no need to support old clang-format tool version as the developers
working on the top of the tree. Further, CI is fragile in this matter.
Furthermore, on macOS dev platform, satisfying the CI's needs natively
and locally has been infeasible.
See
https://gitlab.com/stavallo/ns-3-unapl/-/merge_requests/473#note_2743401748

Improvement:
Support clang-format-19 only

Interop with other CI, such as in ns-3-dev:
For those who desires to write once, pass multiple CIs with different
rules of formatting checker, here are suggestions.
- Writing once posting multiple times was anyway broken already due to
various consideration factors, including confidentiality, timeliness,
and SI-Units. So this is not a new issue.
- One should be able to continue to use style checker/fixer utility
of the other projects. No new impedance.